Move tests to trait files
